Frontend/app/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.9.0",
    "firebase-admin>=7.1.0",
    "google-cloud-pubsub>=2.33.0",
]
